refactor: per-page canonicals and locale detection
