+++ app/resume/page.tsx
@@ -3,7 +3,8 @@ import { Separator } from "@/components/ui/separator"
 import { MapPin, Linkedin, Github } from "lucide-react"
 import { getSettings } from "@/lib/settings"
 import type { Metadata } from "next"
-import { cookies } from "next/headers"
+import { getCanonicalUrl } from "@/utils/getCanonicalUrl"
+import { getLocaleFromPath } from "@/utils/getLocaleFromPath"
 import fs from "fs/promises"
 import path from "path"
 import { marked } from "marked"
@@ -13,8 +14,7 @@ export const revalidate = 60 * 60 * 24
 
 export async function generateMetadata(): Promise<Metadata> {
   const { siteName } = getSettings()
-  const cookieStore = cookies()
-  const locale = (cookieStore.get("NEXT_LOCALE")?.value as "en" | "es") || "en"
+  const locale = getLocaleFromPath()
   const dirPath = path.join(process.cwd(), "public", locale, "resume")
 
   let description = ""
@@ -46,12 +46,16 @@ export async function generateMetadata(): Promise<Metadata> {
 
   const title = siteName
   const image = "/linkedin-picture.jpeg"
+  const siteUrl = getCanonicalUrl()
+  const url = `${siteUrl}${locale === "es" ? "/es" : ""}/resume`
   return {
     title,
     description,
+    alternates: { canonical: url },
     openGraph: {
       title,
       description,
+      url,
       images: [
         {
           url: image,
@@ -72,8 +76,7 @@ export async function generateMetadata(): Promise<Metadata> {
 
 export default async function ResumePage() {
   const { siteName: name } = getSettings()
-  const cookieStore = cookies()
-  const locale = (cookieStore.get("NEXT_LOCALE")?.value as "en" | "es") || "en"
+  const locale = getLocaleFromPath()
   const dirPath = path.join(process.cwd(), "public", locale, "resume")
 
   let sections: { title: string; order: number; html: string }[] = []
