+++ app/sitemap.ts
@@ -8,57 +8,24 @@ import { getCanonicalUrl } from "@/utils/getCanonicalUrl"
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   const siteUrl = getCanonicalUrl()
-  const routes: MetadataRoute.Sitemap = [
-    {
-      url: siteUrl,
-      lastModified: new Date(),
-      alternates: { languages: { es: `${siteUrl}/es` } },
-    },
-    {
-      url: `${siteUrl}/blog`,
-      lastModified: new Date(),
-      alternates: { languages: { es: `${siteUrl}/es/blog` } },
-    },
-    {
-      url: `${siteUrl}/projects`,
-      lastModified: new Date(),
-      alternates: { languages: { es: `${siteUrl}/es/projects` } },
-    },
-    {
-      url: `${siteUrl}/resume`,
-      lastModified: new Date(),
-      alternates: { languages: { es: `${siteUrl}/es/resume` } },
-    },
-    {
-      url: `${siteUrl}/digital-garden`,
-      lastModified: new Date(),
-      alternates: { languages: { es: `${siteUrl}/es/digital-garden` } },
-    },
-  ]
+  const now = new Date()
+  const staticPaths = ["", "/blog", "/projects", "/resume", "/digital-garden"]
+  const routes: MetadataRoute.Sitemap = []
 
-  routes.push(
-    { url: `${siteUrl}/es`, lastModified: new Date(), alternates: { languages: { en: siteUrl } } },
-    {
-      url: `${siteUrl}/es/blog`,
-      lastModified: new Date(),
-      alternates: { languages: { en: `${siteUrl}/blog` } },
-    },
-    {
-      url: `${siteUrl}/es/projects`,
-      lastModified: new Date(),
-      alternates: { languages: { en: `${siteUrl}/projects` } },
-    },
-    {
-      url: `${siteUrl}/es/resume`,
-      lastModified: new Date(),
-      alternates: { languages: { en: `${siteUrl}/resume` } },
-    },
-    {
-      url: `${siteUrl}/es/digital-garden`,
-      lastModified: new Date(),
-      alternates: { languages: { en: `${siteUrl}/digital-garden` } },
-    },
-  )
+  staticPaths.forEach((p) => {
+    const enUrl = `${siteUrl}${p}`
+    const esUrl = `${siteUrl}/es${p}`
+    routes.push({
+      url: enUrl,
+      lastModified: now,
+      alternates: { languages: { es: esUrl } },
+    })
+    routes.push({
+      url: esUrl,
+      lastModified: now,
+      alternates: { languages: { en: enUrl } },
+    })
+  })
 
   try {
     const settings = getNostrSettings()
